refactor(be): split oversized game services under 500-line limit
- Extract handleHistoryNote to games.gateway.history-note.ts
- Extract stripDisabledRules to games.service-rules.ts
- Extract chat methods to game-rooms.chat.service.ts
- Extract generateInviteCode to game-rooms.invite-code.ts
- Extract state size check to game-sessions.size-check.ts
