+++ apps/be/src/games/rooms/game-rooms.invite-code.ts
@@ -1,0 +1,16 @@
+import { randomBytes } from 'crypto';
+import type { Model } from 'mongoose';
+import type { GameRoom } from '../schemas/game-room.schema';
+
+export async function generateUniqueInviteCode(
+  roomModel: Model<GameRoom>,
+): Promise<string> {
+  let code: string;
+  let exists = true;
+  while (exists) {
+    code = randomBytes(4).toString('hex').toUpperCase();
+    const existing = await roomModel.findOne({ inviteCode: code }).exec();
+    exists = !!existing;
+  }
+  return code!;
+}
