+++ apps/be/src/games/sessions/game-sessions.service.ts
@@ -16,6 +16,9 @@ import {
   BaseGameState,
 } from '../engines/base/game-engine.interface';
 import { OCI_CONNECTION } from '../../common/providers/mongo-connections.provider';
+import {
+  enforceStateSizeLimit,
+} from './game-sessions.size-check';
 
 export interface GameSessionSummary {
   id: string;
@@ -52,9 +55,10 @@ export interface ExecuteActionOptions {
  * Game Sessions Service
  * Handles game session lifecycle and state management
  */
-/** Max session document size in bytes. Typical: 2-13KB. Alert at 100KB, strip at 500KB. */
-const WARN_DOC_SIZE_BYTES = 100 * 1024;
-const STRIP_DOC_SIZE_BYTES = 500 * 1024;
+/**
+ * Game Sessions Service
+ * Handles game session lifecycle and state management
+ */
 
 @Injectable()
 export class GameSessionsService {
@@ -189,23 +193,7 @@ export class GameSessionsService {
     }
 
     // Safety valve: strip stateHistory if document is approaching BSON limit
-    const approxSize = Buffer.byteLength(
-      JSON.stringify(session.state),
-      'utf-8',
-    );
-    if (approxSize > STRIP_DOC_SIZE_BYTES) {
-      this.logger.warn(
-        `Session ${sessionId} state is ${Math.round(approxSize / 1024)}KB — stripping stateHistory and logs.`,
-      );
-      const s = session.state;
-      if (Array.isArray(s.stateHistory)) s.stateHistory = [];
-      if (Array.isArray(s.logs)) s.logs = s.logs.slice(-20);
-      session.markModified('state');
-    } else if (approxSize > WARN_DOC_SIZE_BYTES) {
-      this.logger.warn(
-        `Session ${sessionId} state is ${Math.round(approxSize / 1024)}KB — approaching size limit.`,
-      );
-    }
+    enforceStateSizeLimit(session, sessionId, this.logger);
 
     session.updatedAt = new Date();
 
@@ -311,23 +299,7 @@ export class GameSessionsService {
       }
 
       // Safety valve: strip stateHistory if document is approaching BSON limit
-      const approxSize = Buffer.byteLength(
-        JSON.stringify(session.state),
-        'utf-8',
-      );
-      if (approxSize > STRIP_DOC_SIZE_BYTES) {
-        this.logger.warn(
-          `Session ${sessionId} state is ${Math.round(approxSize / 1024)}KB — stripping stateHistory and logs.`,
-        );
-        const s = session.state;
-        if (Array.isArray(s.stateHistory)) s.stateHistory = [];
-        if (Array.isArray(s.logs)) s.logs = s.logs.slice(-20);
-        session.markModified('state');
-      } else if (approxSize > WARN_DOC_SIZE_BYTES) {
-        this.logger.warn(
-          `Session ${sessionId} state is ${Math.round(approxSize / 1024)}KB — approaching size limit.`,
-        );
-      }
+      enforceStateSizeLimit(session, sessionId, this.logger);
 
       session.updatedAt = new Date();
